refactor: App.vue에서 분리된 전역 트랜지션 제어 및 테마 세팅 로직
