+++ DUNOESANCHAEG-Client/src/App.vue
@@ -1,36 +1,31 @@
 <!-- 레이아웃이 필요 없는 페이지는 meta: { hideLayout: true }를 추가 -->
 <template>
-  <template v-if="route.meta.hideLayout">
-    <router-view />
-  </template>
-
-  <MainLayout v-else>
-    <router-view />
-  </MainLayout>
+  <router-view v-slot="{ Component, route }">
+    <transition :name="transitionName" mode="out-in">
+      <div :key="route.meta.hideLayout ? 'no-layout' : 'has-layout'" class="w-full min-h-dvh">
+        <component :is="Component" v-if="route.meta.hideLayout" />
+        
+        <MainLayout v-else>
+          <component :is="Component" />
+        </MainLayout>
+      </div>
+    </transition>
+  </router-view>
 </template>
 
 <script setup>
-import { onMounted, watch } from 'vue';
 import { useRoute } from 'vue-router';
 import MainLayout from './layouts/MainLayout.vue';
-import { useAuthStore } from '@/store/auth';
-import { useSettingsStore } from '@/store/settings';
+import { useRouteTransition } from '@/composables/useRouteTransition';
+import { useThemeSettings } from '@/composables/useThemeSettings';
 
 const route = useRoute();
-const authStore = useAuthStore();
-const settingsStore = useSettingsStore();
 
-onMounted(() => {
-  settingsStore.initSettings(authStore.fontSize, authStore.isHighContrast);
-});
+// 글로벌 테마 및 설정 관찰 로직
+useThemeSettings();
 
-watch(
-    () => [authStore.isHighContrast, authStore.fontSize],
-    ([newContrast, newSize]) => {
-      settingsStore.initSettings(newSize, newContrast);
-    }
-);
-
+// 동적 슬라이드 트랜지션 애니메이션 로직
+const { transitionName } = useRouteTransition();
 </script>
 
 <style>
